+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Add --no-prompt option to help.
 
+### Fixed
+- `install` command.
+
 ### Removed
 - `plugin *` commands.